[MDH-15] refactor : waiting 서비스 메서드 분리
