+++ src/main/java/com/mdh/devtable/waiting/application/WaitingService.java
@@ -1,9 +1,9 @@
 package com.mdh.devtable.waiting.application;
 
-import com.mdh.devtable.waiting.Waiting;
-import com.mdh.devtable.waiting.WaitingLine;
-import com.mdh.devtable.waiting.WaitingPeople;
+import com.mdh.devtable.waiting.domain.Waiting;
+import com.mdh.devtable.waiting.domain.WaitingPeople;
 import com.mdh.devtable.waiting.infra.persistence.ShopWaitingRepository;
+import com.mdh.devtable.waiting.infra.persistence.WaitingLine;
 import com.mdh.devtable.waiting.infra.persistence.WaitingRepository;
 import com.mdh.devtable.waiting.presentation.dto.WaitingCreateRequest;
 import lombok.RequiredArgsConstructor;
@@ -24,9 +24,7 @@ public class WaitingService {
         var shopWaiting = shopWaitingRepository.findById(shopId)
             .orElseThrow(() -> new IllegalStateException("해당 매장에 웨이팅 정보가 존재하지 않습니다. shopId : " + shopId));
 
-        var adultCount = waitingCreateRequest.adultCount();
-        var childCount = waitingCreateRequest.childCount();
-        var waitingPeople = new WaitingPeople(adultCount, childCount);
+        WaitingPeople waitingPeople = createWaitingPeople(waitingCreateRequest);
 
         var userId = waitingCreateRequest.userId();
         var waiting = Waiting.builder()
@@ -35,11 +33,21 @@ public class WaitingService {
             .userId(userId)
             .build();
 
-        var savedWaiting = waitingRepository.save(waiting);
+        var savedWaiting = waitingRepository.save(waiting); // 저장
+        saveWaitingLine(shopId, savedWaiting);
+
+        return savedWaiting.getId();
+    }
+
+    private void saveWaitingLine(Long shopId, Waiting savedWaiting) {
         var waitingId = savedWaiting.getId();
         var createdDate = savedWaiting.getCreatedDate();
-        waitingLine.save(shopId, waitingId, createdDate);
+        waitingLine.save(shopId, waitingId, createdDate); // 웨이팅 라인 저장
+    }
 
-        return waitingId;
+    private WaitingPeople createWaitingPeople(WaitingCreateRequest waitingCreateRequest) {
+        var adultCount = waitingCreateRequest.adultCount();
+        var childCount = waitingCreateRequest.childCount();
+        return new WaitingPeople(adultCount, childCount);
     }
 }
